Непрокручиваемый поиск, Смена заднего фона при прогрузке, рефакторинг
