+++ lib/feature/presentation/ui/contacts_page/contact_profile_pop_up.dart
@@ -2,7 +2,8 @@ import 'package:cportal_flutter/common/custom_theme.dart';
 import 'package:cportal_flutter/feature/presentation/bloc/get_single_profile_bloc/get_single_profile_bloc.dart';
 import 'package:cportal_flutter/feature/presentation/bloc/get_single_profile_bloc/get_single_profile_event.dart';
 import 'package:cportal_flutter/feature/presentation/bloc/get_single_profile_bloc/get_single_profile_state.dart';
-import 'package:cportal_flutter/feature/presentation/ui/contacts_page/widgets/profile_info_section.dart';
+import 'package:cportal_flutter/feature/presentation/ui/contacts_page/widgets/profile_image.dart';
+import 'package:cportal_flutter/feature/presentation/ui/widgets/profile_info_section.dart';
 import 'package:cportal_flutter/feature/presentation/ui/contacts_page/widgets/contacts_list.dart';
 import 'package:cportal_flutter/feature/presentation/ui/widgets/avatar_box.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
@@ -76,20 +77,7 @@ class _ContactProfilePopUpState extends State<ContactProfilePopUp> {
                         crossAxisAlignment: CrossAxisAlignment.start,
                         children: [
                           // Profile Image & Full Name.
-                          Align(
-                            alignment: Alignment.center,
-                            child: state.profile.photoLink != ''
-                                ? AvatarBox(
-                                    size: 102,
-                                    imgPath: state.profile.photoLink,
-                                    borderRadius: 24,
-                                  )
-                                : EmptyAvatarBox(
-                                    size: 102,
-                                    borderRadius: 24,
-                                    user: state.profile,
-                                  ),
-                          ),
+                          ProfileImage(user: state.profile),
                           const SizedBox(height: 12),
                           Align(
                             alignment: Alignment.center,
@@ -109,14 +97,14 @@ class _ContactProfilePopUpState extends State<ContactProfilePopUp> {
                             text: state.profile.position,
                             bottomPadding: 18,
                           ),
-          
+
                           // Department.
                           ProfileInfoSection(
                             headline: AppLocalizations.of(context)!.department,
                             text: state.profile.department,
                             bottomPadding: 18,
                           ),
-          
+
                           // Contact info.
                           ...List.generate(
                             state.profile.contactInfo.length,
@@ -126,11 +114,12 @@ class _ContactProfilePopUpState extends State<ContactProfilePopUp> {
                               bottomPadding: 18,
                             ),
                           ),
-          
+
                           // Birth date.
                           if (state.profile.birthDayToString != null)
                             ProfileInfoSection(
-                              headline: AppLocalizations.of(context)!.birth_date,
+                              headline:
+                                  AppLocalizations.of(context)!.birth_date,
                               text: state.profile.birthDayToString!,
                               bottomPadding: 0,
                             ),
